requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
psycopg2-binary==2.8.5
pytz==2020.1
Shapely==1.7.0
sqlparse==0.3.1
